# web/docs/pdf_sku_openapi_v2_0.yaml
openapi: 3.0.3
info:
  title: PDF-SKU Extraction System API
  description: |
    PDF 自动分类与 SKU 提取系统 — 完整 API 契约

    **架构**: 模块化单体（Modular Monolith）  
    **认证**: Bearer Token (JWT)  
    **对齐文档**: TA V1.7 §5 · UI/UX V1.4 §8 · 接口契约总表 V1.1 · BA V1.1 · 后端详设 V1.2 · LLM Adapter V1.2 · 前端详设 V1.2

    ## 状态枚举冻结声明
    ⚠ Job/Page/SKU 状态枚举为冻结版本 V2.0。变更需前后端同步发布，禁止单端修改。

    ## V2.0 合并说明
    以 V1.2（对齐模块设计）为主体，精选吸收 V1.7 的 15 项增强：
    - [V1.7→] JobDetail 双轨状态 user_status + action_hint + 4 时间戳
    - [V1.7→] TaskDetail 增加 locked_by/locked_at/rework_count/timeout_at
    - [V1.7→] ImageBinding 丰富字段 (binding_method/is_duplicate/image_hash)
    - [V1.7→] POST /annotations 独立标注端点 (8 种 type)
    - [V1.7→] GET /metrics · GET /ops/dashboard/events · POST /ops/tasks/batch-skip
    - [V1.7→] /ops/custom-attr-upgrades (GET + POST)
    - [V1.7→] GET /jobs 增加 created_after/created_before
    - [V1.7→] OPTIONS /uploads (tus CORS)
    - [V1.7→] Evaluation 增加 route_reason/sampling/prompt_version
    - [V1.7→] PrescanResult 增加 raw_metrics
    - [V1.7→] Tag 细分 (12 tag) + PaginationMeta 独立 schema
    - [V2.1→] Auth 模块 9 个端点（login/register/me/change-password/admin CRUD）
    - [V2.1→] User schema + LoginRequest/LoginResponse/RegisterRequest

    ## V0.2.0 新增端点
    - GET /jobs/{job_id}/pages/{page_no}/detail — 页面+SKU+图片合并响应
    - GET /jobs/{job_id}/images/{image_id} — 提取的产品图片服务

    **未吸收 V1.7 项**（与模块设计不兼容）：
    - TaskCompleteRequest skus 模型 → 保留 V1.2 的 groups 元素-分组模型
...
    - ErrorResponse 去掉 severity → 保留
...
servers:
...
security:
  - BearerAuth: []

# ====================================================================
# Tags (V1.7 细分 12 tag)
# ====================================================================
...
paths:
...
  /tasks:
    get:
      tags: [Tasks]
      summary: 待标注任务列表
      parameters:
        - in: query
          name: group_by
          schema: { type: string, enum: [file] }
          description: 按文件聚合
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items: { $ref: '#/components/schemas/TaskFileGroup' }
  /tasks/next:
    post:
      tags: [Tasks]
      summary: 自动领取下一个优先级最高的任务
      description: |
        服务端按 SLA 优先级排序，选择队列头部任务自动 lock。
        等价于 GET /tasks 取第一条 + POST /tasks/{id}/lock 的原子操作。
        如无可用任务返回 204。
      responses:
        '200':
          description: 已领取
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskDetail' }
        '204':
          description: 暂无待领取任务
        '409':
          description: 并发冲突（其他标注员抢先领取），客户端应重试
  /tasks/{task_id}:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    get:
      tags: [Tasks]
      summary: 任务详情（含 AI 元素 + 歧义绑定）
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskDetail' }
...
  /tasks/{task_id}/heartbeat:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    post:
      tags: [Tasks]
      summary: 锁续期心跳（每 30s）
      responses:
        '200':
          description: 续期成功
        '409':
          description: 锁已过期
...
  /tasks/{task_id}/complete:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    post:
      tags: [Tasks]
      summary: 提交标注并完成任务（原子操作）
      description: |
        请求体包含完整标注数据，使用元素-分组（element-group）模型。
        对齐 UI/UX §6.5 套索→分组→填属性→提交流程。
        对齐后端 V1.1 Collaboration 模块 task_complete handler。
        对齐前端 V1.0 annotationStore.buildSubmitPayload()。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskCompletePayload' }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  extracted_sku_count: { type: integer }
                  imported_count: { type: integer }
        '400':
          description: 标注数据校验失败
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ValidationError' }
        '409':
          description: 锁已过期或任务已被重新分配
...
  /tasks/{task_id}/revert:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    post:
      tags: [Tasks]
      summary: 撤销已完成任务（组长权限）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        '200':
          description: 已撤销
        '409':
          description: 任务不可撤销

  # ======================== Annotations (V1.7→) ========================
  /annotations:
    post:
      tags: [Annotations]
      summary: 提交标注记录（独立于任务提交）
      description: |
        8 种标注类型（BA §4.8）：PAGE_TYPE_CORRECTION、TEXT_ROLE_CORRECTION、
        IMAGE_ROLE_CORRECTION、SKU_ATTRIBUTE_CORRECTION、BINDING_CORRECTION、
        CUSTOM_ATTR_CONFIRM、NEW_TYPE_REPORT、LAYOUT_CORRECTION。

        此端点记录标注员的修正行为，供 Feedback 模块收集用于模型优化。
        与 /tasks/{id}/complete 不同：complete 是提交最终结果，此处是记录过程中的修正。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAnnotationRequest' }
      responses:
        '201':
          description: 标注记录已保存
          content:
            application/json:
              schema:
                type: object
                properties:
                  annotation_id: { type: string, format: uuid }
...
  /ops/tasks/batch-reassign:
    post:
      tags: [Ops-Batch]
      summary: 批量重分配标注任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                task_ids: { type: array, items: { type: string, format: uuid } }
                target_annotator_id: { type: string }
      responses:
        '200':
          description: 已重分配
  /ops/tasks/batch-skip:
    post:
      tags: [Ops-Batch]
      summary: 批量跳过任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task_ids]
              properties:
                task_ids: { type: array, items: { type: string, format: uuid } }
                reason: { type: string }
      responses:
        '200':
          description: 操作完成
          content:
            application/json:
              schema:
                type: object
                properties:
                  success_count: { type: integer }
                  failed_count: { type: integer }

  # ======================== Ops-Config ========================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Token 由 POST /auth/login 签发，有效期 24h。
        实现: python-jose + pbkdf2_hmac 密码哈希。
        环境变量: JWT_SECRET_KEY（必须配置）
        Header: `Authorization: Bearer <token>`
        Payload: { user_id, username, role, merchant_id?, exp }
  parameters:
...
    TaskId:
      in: path
      name: task_id
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    TaskFileGroup:
      type: object
      properties:
        job_id: { type: string, format: uuid }
        source_file: { type: string }
        merchant_id: { type: string }
        pending_count: { type: integer }
        total_count: { type: integer }
    TaskDetail:
      type: object
      description: 合并 V1.2 业务字段 + V1.7 锁状态可观测字段
      properties:
        task_id: { type: string, format: uuid }
        job_id: { type: string, format: uuid }
        page_number: { type: integer }
        task_type: { type: string }
        status: { type: string }
        priority: { type: string, enum: [NORMAL, HIGH, URGENT, CRITICAL, AUTO_RESOLVE] }
        sla_deadline: { type: string, format: date-time, nullable: true }
        sla_level: { type: string, enum: [NORMAL, HIGH, CRITICAL, AUTO_RESOLVE] }
        # V1.7→ 锁状态可观测
        locked_by: { type: string, nullable: true, description: 当前锁定者 (V1.7→) }
        locked_at: { type: string, format: date-time, nullable: true }
        timeout_at: { type: string, format: date-time, nullable: true, description: 锁过期时间 }
        assigned_to: { type: string, nullable: true }
        assigned_at: { type: string, format: date-time, nullable: true }
        rework_count: { type: integer, default: 0, description: 返工次数 (V1.7→) }
        created_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }
        # 业务上下文
        context:
          type: object
          properties:
            page_type: { type: string }
            layout_type: { type: string }
            screenshot_url: { type: string }
            ai_result: { type: object }
            cross_page_table: { type: object, nullable: true }
        elements:
          type: array
          items: { $ref: '#/components/schemas/AnnotationElement' }
        ambiguous_bindings:
          type: array
          items: { $ref: '#/components/schemas/AmbiguousBinding' }
    AnnotationElement:
      type: object
      properties:
        element_id: { type: string }
        type: { type: string, enum: [image, text] }
        bbox:
          type: object
          properties:
            x: { type: number, minimum: 0, maximum: 1 }
            y: { type: number, minimum: 0, maximum: 1 }
            w: { type: number, minimum: 0, maximum: 1 }
            h: { type: number, minimum: 0, maximum: 1 }
        ai_role: { type: string }
        confidence: { type: number, format: float }
    AmbiguousBinding:
      type: object
      description: 歧义绑定（top1-top2 < 0.2），对齐 TA ADR-14
      properties:
        element_id: { type: string }
        candidates:
          type: array
          items:
            type: object
            properties:
              image_uri: { type: string }
              confidence: { type: number }
              rank: { type: integer }
        resolved: { type: boolean }
        selected_rank: { type: integer, nullable: true }

    # ======== Task Submit (V1.2 元素-分组模型，对齐前端/后端详设) ========
    TaskCompletePayload:
      type: object
      description: |
        标注提交 payload — 元素-分组（element-group）模型。
        对齐: UI/UX §6.5 · 前端 V1.0 annotationStore · 后端 V1.1 Collaboration
      required: [task_id, page_type, groups]
      properties:
        task_id: { type: string, format: uuid }
        page_type: { type: string, enum: [A, B, C, D] }
        layout_type: { type: string, enum: [L1, L2, L3, L4] }
        groups:
          type: array
          items:
            type: object
            required: [group_id, sku_type, elements]
            properties:
              group_id: { type: string }
              label: { type: string }
              sku_type: { type: string, enum: [complete, partial, invalid] }
              elements:
                type: array
                items: { $ref: '#/components/schemas/AnnotationElement' }
              sku_attributes:
                type: object
                properties:
                  model: { type: string }
                  name: { type: string }
                  size: { type: string }
                  material: { type: string }
                  color: { type: string }
                  price: { type: string }
              custom_attributes:
                type: array
                items:
                  type: object
                  properties:
                    key: { type: string }
                    value: { type: string }
              partial_contains:
                type: array
                items: { type: string }
              cross_page_sku_id: { type: string, nullable: true }
              invalid_reason: { type: string, nullable: true }
        ungrouped_elements:
          type: array
          items: { type: string }
        binding_confirmations:
          type: array
          description: 歧义绑定确认结果
          items:
            type: object
            properties:
              element_id: { type: string }
              selected_rank: { type: integer }
        feedback:
          type: object
          properties:
            page_type_modified: { type: boolean }
            layout_type_modified: { type: boolean }
            new_image_role_observed: { type: boolean }
            new_text_role_observed: { type: boolean }
            notes: { type: string }
...
    CreateAnnotationRequest:
      type: object
      description: |
        8 种标注类型（BA §4.8），独立于 TaskCompletePayload。
        记录标注过程中的修正行为，供 Feedback 模块收集。
      required: [job_id, page_number, type, payload]
      properties:
        task_id: { type: string, format: uuid, nullable: true }
        job_id: { type: string, format: uuid }
        page_number: { type: integer }
        type:
          type: string
          enum:
            - PAGE_TYPE_CORRECTION
            - TEXT_ROLE_CORRECTION
            - IMAGE_ROLE_CORRECTION
            - SKU_ATTRIBUTE_CORRECTION
            - BINDING_CORRECTION
            - CUSTOM_ATTR_CONFIRM
            - NEW_TYPE_REPORT
            - LAYOUT_CORRECTION
        payload:
          type: object
          additionalProperties: true
          description: 按 type 区分的结构化内容（BA §4.8 Payload Schema）

    # ======== SSE Events ========
...
    ValidationError:
      type: object
      properties:
        code: { type: string, example: ANNOTATION_VALIDATION_FAILED }
        message: { type: string }
        field_errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }